* Optimize 3-arg (de)interleave from successive indexes with AVX * New load_interleaved and store_interleaved functions for simdized   objects. This uses existing interleave code, when the sizeof all   members is equal (homogeneous_sizeof). Otherwise it'll use loads from   scalars and depend on the compiler to do something smarter. * Modify simd_for_each(_n) to use simdize<value_type> and thus vectorize   everything that's thrown at it. This uses the new load_interleaved and   store_interleaved functions to improve the chunked AoS <-> AoVS   conversions.

+++ Vc/scalar/detail.h
@@ -35,7 +35,7 @@ namespace Vc_VERSIONED_NAMESPACE
 namespace Detail
 {
 //InterleaveImpl{{{1
-template<typename V, int Size, size_t VSize> struct InterleaveImpl;
+template<typename V, size_t Size, size_t VSize> struct InterleaveImpl;
 template<typename V, size_t VSize> struct InterleaveImpl<V, 1, VSize> {
     template <typename I>  // interleave 2 args {{{2
     static inline void interleave(typename V::EntryType *const data, const I &i,
